test: write basic test for Product
