+++ db/models/review.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const {TEXT} = require('sequelize')
+const { TEXT } = require('sequelize')
 
 module.exports = db => db.define('reviews', {
   body: {
@@ -18,5 +18,3 @@ module.exports.associations = (Review, {Product, User}) => {
   Review.belongsTo(Product)
   Review.belongsTo(User)
 }
-
-// or this can be in Product / User models where Product.hasMany(Review) and User.hasMany(Review)
