+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "TENKShopper",
+  "name": "tenk-shopper",
   "version": "0.0.1",
   "description": "The best Grace Shopper in all the land.",
   "main": "index.js",
@@ -19,7 +19,7 @@
     "start": "node server/start.js",
     "start-watch": "nodemon server/start.js --watch server --watch db --watch index.js --watch package.json",
     "start-dev": "cross-env NODE_ENV=development npm run start-watch",
-    "test": "mocha --compilers js:babel-register --watch-extensions js,jsx app/**/*.test.js app/**/*.test.jsx server/**/*.test.js db/**/*.test.js",
+    "test": "mocha --compilers js:babel-register --watch-extensions js,jsx test/*.test.js app/**/*.test.js app/**/*.test.jsx server/**/*.test.js db/**/*.test.js",
     "test-watch": "npm run test -- --watch --reporter=min",
     "seed": "node db/seed.js",
     "deploy-heroku": "bin/deploy-heroku.sh",
